refactor: split gml.rb into individual files with shared attribute modules
Extract HasBaseClass and HasCollectionProperties modules to DRY up
the repeated isCollection/noPropertyType/base_Class pattern across
DataType, Type, Union, FeatureType, and CodeList. Split 133-line
gml.rb into 9 one-class-per-file modules with autoload.
