+++ lib/xmi/sparx/gml/union.rb
@@ -1,0 +1,18 @@
+# frozen_string_literal: true
+
+module Xmi
+  module Sparx
+    module Gml
+      class Union < Lutaml::Model::Serializable
+        include HasCollectionProperties
+
+        xml do
+          root "Union"
+          namespace ::Xmi::Namespace::Sparx::Gml
+
+          HasCollectionProperties.apply_xml_mappings(self)
+        end
+      end
+    end
+  end
+end
